rename error package
