+++ src/main/java/com/teddycrane/springpractice/error/BadRequestException.java
@@ -1,4 +1,4 @@
-package com.teddycrane.springpractice.exceptions;
+package com.teddycrane.springpractice.error;
 
 public class BadRequestException extends BaseRequestException{
 	public BadRequestException(String message) {
