+++ src/main/java/com/teddycrane/springpractice/error/CreationException.java
@@ -1,0 +1,7 @@
+package com.teddycrane.springpractice.error;
+
+public class CreationException extends BaseRequestException {
+	public CreationException(String message) {
+		super(message);
+	}
+}
