+++ src/main/java/com/teddycrane/springpractice/error/RaceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.teddycrane.springpractice.exceptions;
+package com.teddycrane.springpractice.error;
 
 public class RaceNotFoundException extends BaseNotFoundException{
 	public RaceNotFoundException(String message) {
