+++ src/main/java/com/teddycrane/springpractice/error/UpdateException.java
@@ -1,0 +1,7 @@
+package com.teddycrane.springpractice.error;
+
+public class UpdateException extends BaseRequestException {
+	public UpdateException(String message) {
+		super(message);
+	}
+}
