+++ src/main/java/com/teddycrane/springpractice/error/UserNotFoundError.java
@@ -1,4 +1,4 @@
-package com.teddycrane.springpractice.exceptions;
+package com.teddycrane.springpractice.error;
 
 public class UserNotFoundError extends BaseNotFoundException
 {
