+++ src/main/java/com/teddycrane/springpractice/event/model/IEventController.java
@@ -3,9 +3,9 @@ package com.teddycrane.springpractice.event.model;
 import com.teddycrane.springpractice.event.request.CreateEventRequest;
 import com.teddycrane.springpractice.event.Event;
 import com.teddycrane.springpractice.event.request.UpdateEventRequest;
-import com.teddycrane.springpractice.exceptions.BadRequestException;
-import com.teddycrane.springpractice.exceptions.DuplicateItemException;
-import com.teddycrane.springpractice.exceptions.EventNotFoundException;
+import com.teddycrane.springpractice.error.BadRequestException;
+import com.teddycrane.springpractice.error.DuplicateItemException;
+import com.teddycrane.springpractice.error.EventNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
