--- src/main/java/com/teddycrane/springpractice/exceptions/CreationException.java
@@ -1,10 +1,0 @@
-package com.teddycrane.springpractice.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-
-public class CreationException extends BaseRequestException {
-	public CreationException(String message) {
-		super(message);
-	}
-}
