--- src/main/java/com/teddycrane/springpractice/exceptions/UpdateException.java
@@ -1,10 +1,0 @@
-package com.teddycrane.springpractice.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-
-public class UpdateException extends BaseRequestException {
-	public UpdateException(String message) {
-		super(message);
-	}
-}
