+++ src/main/java/com/teddycrane/springpractice/race/RaceService.java
@@ -5,7 +5,7 @@ import com.teddycrane.springpractice.race.model.RaceResult;
 import com.teddycrane.springpractice.racer.Racer;
 import com.teddycrane.springpractice.enums.Category;
 import com.teddycrane.springpractice.enums.RaceFilterType;
-import com.teddycrane.springpractice.exceptions.*;
+import com.teddycrane.springpractice.error.*;
 import com.teddycrane.springpractice.helper.EnumHelpers;
 import com.teddycrane.springpractice.models.Either;
 import com.teddycrane.springpractice.race.model.RaceRepository;
