+++ src/main/java/com/teddycrane/springpractice/racer/RacerController.java
@@ -3,9 +3,9 @@ package com.teddycrane.springpractice.racer;
 import com.teddycrane.springpractice.models.BaseController;
 import com.teddycrane.springpractice.enums.Category;
 import com.teddycrane.springpractice.enums.RacerFilterType;
-import com.teddycrane.springpractice.exceptions.BadRequestException;
-import com.teddycrane.springpractice.exceptions.RacerNotFoundException;
-import com.teddycrane.springpractice.exceptions.UpdateException;
+import com.teddycrane.springpractice.error.BadRequestException;
+import com.teddycrane.springpractice.error.RacerNotFoundException;
+import com.teddycrane.springpractice.error.UpdateException;
 import com.teddycrane.springpractice.helper.EnumHelpers;
 import com.teddycrane.springpractice.racer.model.IRacerController;
 import com.teddycrane.springpractice.racer.model.IRacerService;
