+++ src/main/java/com/teddycrane/springpractice/racer/model/IRacerController.java
@@ -1,7 +1,7 @@
 package com.teddycrane.springpractice.racer.model;
 
-import com.teddycrane.springpractice.exceptions.BadRequestException;
-import com.teddycrane.springpractice.exceptions.RacerNotFoundException;
+import com.teddycrane.springpractice.error.BadRequestException;
+import com.teddycrane.springpractice.error.RacerNotFoundException;
 import com.teddycrane.springpractice.racer.request.CreateRacerRequest;
 import com.teddycrane.springpractice.racer.Racer;
 import com.teddycrane.springpractice.racer.request.UpdateRacerRequest;
