+++ src/test/java/com/teddycrane/springpractice/tests/controllertests/RaceControllerTest.java
@@ -8,10 +8,10 @@ import com.teddycrane.springpractice.race.request.UpdateRaceRequest;
 import com.teddycrane.springpractice.racer.request.AddRacerRequest;
 import com.teddycrane.springpractice.racer.Racer;
 import com.teddycrane.springpractice.enums.Category;
-import com.teddycrane.springpractice.exceptions.BadRequestException;
-import com.teddycrane.springpractice.exceptions.DuplicateItemException;
-import com.teddycrane.springpractice.exceptions.RaceNotFoundException;
-import com.teddycrane.springpractice.exceptions.UpdateException;
+import com.teddycrane.springpractice.error.BadRequestException;
+import com.teddycrane.springpractice.error.DuplicateItemException;
+import com.teddycrane.springpractice.error.RaceNotFoundException;
+import com.teddycrane.springpractice.error.UpdateException;
 import com.teddycrane.springpractice.race.model.IRaceService;
 import com.teddycrane.springpractice.tests.helpers.TestResourceGenerator;
 import org.junit.Assert;
