+++ src/test/java/com/teddycrane/springpractice/tests/controllertests/RacerControllerTest.java
@@ -4,8 +4,8 @@ import com.teddycrane.springpractice.racer.RacerController;
 import com.teddycrane.springpractice.racer.Racer;
 import com.teddycrane.springpractice.enums.Category;
 import com.teddycrane.springpractice.enums.RacerFilterType;
-import com.teddycrane.springpractice.exceptions.BadRequestException;
-import com.teddycrane.springpractice.exceptions.RacerNotFoundException;
+import com.teddycrane.springpractice.error.BadRequestException;
+import com.teddycrane.springpractice.error.RacerNotFoundException;
 import com.teddycrane.springpractice.racer.request.CreateRacerRequest;
 import com.teddycrane.springpractice.racer.request.UpdateRacerRequest;
 import com.teddycrane.springpractice.racer.model.IRacerService;
