+++ src/test/java/com/teddycrane/springpractice/tests/servicetests/RacerServiceTest.java
@@ -3,7 +3,7 @@ package com.teddycrane.springpractice.tests.servicetests;
 import com.teddycrane.springpractice.racer.Racer;
 import com.teddycrane.springpractice.enums.Category;
 import com.teddycrane.springpractice.enums.RacerFilterType;
-import com.teddycrane.springpractice.exceptions.RacerNotFoundException;
+import com.teddycrane.springpractice.error.RacerNotFoundException;
 import com.teddycrane.springpractice.racer.model.RacerRepository;
 import com.teddycrane.springpractice.racer.model.IRacerService;
 import com.teddycrane.springpractice.racer.RacerService;
